Firedrake backend: Move projection code
