+++ tlm_adjoint/firedrake/backend_patches.py
@@ -15,11 +15,12 @@ from ..patch import (
     add_manager_controls, manager_method, patch_function, patch_method)
 
 from .backend_interface import Cofunction
-from .equations import Assembly, EquationSolver, Projection
+from .equations import Assembly, EquationSolver
 from .functions import (
     Constant, define_var_alias, expr_zero, extract_coefficients, iter_expr)
-from .firedrake_equations import ExprAssignment, LocalProjection
+from .firedrake_equations import ExprAssignment
 from .interpolation import ExprInterpolation
+from .projection import Projection, LocalProjection
 
 import numbers
 import operator
